+++ Models/LvRefModel.cs
@@ -2,7 +2,7 @@ namespace ServerApp.Controllers
 {
     public class LvRefModel
     {
-        public int TelId { get; set; }
+        public long TelId { get; set; }
         public double? Cena { get; set; }
     }
 }
